Revert "Added more windows and changed packages"
