+++ Application/src/GUIView.java
@@ -1,4 +1,3 @@
-package source;
 /*
  * @projectDescription The main "page" for the JAMBL application
  * 
@@ -6,6 +5,7 @@ package source;
  * 
  * @dateLastModified September 26, 2022
  */
+
 
 import java.awt.EventQueue;
 
@@ -22,15 +22,16 @@ import java.awt.Font;
 // Custom java package containing the pop-up windows
 import PopUps.*;
 
-public class GUIView extends View {
 
+public class GUIView extends View  {
+	
 	Model model = new Model();
 	private Controller GUICntrlr = new Controller(model, this);
-
+	
 	private JFrame frmJambl;
-	// A boolean signifying if the recent UML diagram has been saved.
+	// A boolean signifying if the recent UML diagram has been saved. 
 	// If not, a pop-up message will appear upon clicking the x button
-	// asking the user if they would like to save before closing
+	//     asking the user if they would like to save before closing
 	private boolean saved = false;
 
 	/**
@@ -60,88 +61,87 @@ public class GUIView extends View {
 	 * Initialize the contents of the frame.
 	 */
 	private void initialize() {
-
-		// ******** MAIN PAGE ********//
+		
+		
+		//******** MAIN PAGE ********//
 		frmJambl = new JFrame();
 		frmJambl.setTitle("JAMBL");
 		frmJambl.setBounds(100, 100, 1063, 752);
 		frmJambl.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
 		frmJambl.getContentPane().setLayout(null);
-
-		/////////////////////////////////////////// textAreaMain - where the UML diagram
-		/////////////////////////////////////////// will be able displayed and updated
-		/////////////////////////////////////////// in real time
-		/////////////////////////////////////////// as classes, field, methods, and
-		/////////////////////////////////////////// relationships are added
+		
+		/////////////////////////////////////////// textAreaMain - where the UML diagram will be able displayed and updated in real time
+		///////////////////////////////////////////			as classes, field, methods, and relationships are added
 		JTextArea textAreaMain = new JTextArea();
 		textAreaMain.setEditable(false);
 		textAreaMain.setBounds(359, 67, 678, 635);
 		frmJambl.getContentPane().add(textAreaMain);
-
+		
+		
 		JLabel ClassLabel = new JLabel("Class");
 		ClassLabel.setFont(new Font("Tahoma", Font.PLAIN, 15));
 		ClassLabel.setBounds(10, 68, 214, 13);
 		frmJambl.getContentPane().add(ClassLabel);
-
+		
 		JSeparator separator_1 = new JSeparator();
 		separator_1.setBounds(10, 91, 106, 2);
 		frmJambl.getContentPane().add(separator_1);
-
+		
 		JLabel lblRelationships = new JLabel("Relationships");
 		lblRelationships.setFont(new Font("Tahoma", Font.PLAIN, 15));
 		lblRelationships.setBounds(220, 68, 214, 13);
 		frmJambl.getContentPane().add(lblRelationships);
-
+		
 		JSeparator separator_2 = new JSeparator();
 		separator_2.setBounds(220, 93, 97, 2);
 		frmJambl.getContentPane().add(separator_2);
-
+		
 		/****** TEMPORARY WHILE WHILE WE FIGURE OUT WHAT TO PUT HERE *******/
 		JLabel lblNewLabel = new JLabel("Under Construction!");
 		lblNewLabel.setFont(new Font("Tahoma", Font.BOLD, 15));
 		lblNewLabel.setBounds(10, 16, 187, 13);
 		frmJambl.getContentPane().add(lblNewLabel);
-
+		
 		JLabel lblSaveload = new JLabel("Save/Load");
 		lblSaveload.setFont(new Font("Tahoma", Font.PLAIN, 15));
 		lblSaveload.setBounds(10, 373, 214, 13);
 		frmJambl.getContentPane().add(lblSaveload);
-
+		
 		JSeparator separator_4 = new JSeparator();
 		separator_4.setBounds(10, 398, 106, 2);
 		frmJambl.getContentPane().add(separator_4);
-
+		
 		JLabel lblNewLabel_1 = new JLabel("Parameters");
 		lblNewLabel_1.setFont(new Font("Tahoma", Font.PLAIN, 15));
 		lblNewLabel_1.setBounds(10, 217, 90, 16);
 		frmJambl.getContentPane().add(lblNewLabel_1);
-
+		
 		JSeparator separator_1_1 = new JSeparator();
 		separator_1_1.setBounds(10, 245, 106, 2);
 		frmJambl.getContentPane().add(separator_1_1);
-
+		
 		JLabel lblMethods = new JLabel("Methods");
 		lblMethods.setFont(new Font("Tahoma", Font.PLAIN, 15));
 		lblMethods.setBounds(220, 218, 214, 13);
 		frmJambl.getContentPane().add(lblMethods);
-
+		
 		JSeparator separator_1_1_1 = new JSeparator();
 		separator_1_1_1.setBounds(220, 245, 106, 2);
 		frmJambl.getContentPane().add(separator_1_1_1);
-
+		
 		JLabel lblFields = new JLabel("Fields");
 		lblFields.setFont(new Font("Tahoma", Font.PLAIN, 15));
 		lblFields.setBounds(220, 398, 214, 13);
 		frmJambl.getContentPane().add(lblFields);
-
+		
 		JSeparator separator_4_1 = new JSeparator();
 		separator_4_1.setBounds(220, 423, 106, 2);
 		frmJambl.getContentPane().add(separator_4_1);
-
+		
 		///////////////////////////
-		// ******* Buttons *******//
+		//******* Buttons *******//
 		///////////////////////////
-
+		
 		////////////////////////////////////////////////////////////// Add Class Button
 		JButton btnAddClass = new JButton("Add");
 		btnAddClass.setBounds(10, 103, 97, 21);
@@ -155,9 +155,8 @@ public class GUIView extends View {
 				saved = false;
 			}
 		});
-
-		////////////////////////////////////////////////////////////// Rename Class
-		////////////////////////////////////////////////////////////// Button
+		
+		////////////////////////////////////////////////////////////// Rename Class Button
 		JButton btnRenameClass = new JButton("Rename");
 		btnRenameClass.setBounds(10, 134, 97, 21);
 		frmJambl.getContentPane().add(btnRenameClass);
@@ -170,9 +169,8 @@ public class GUIView extends View {
 				saved = false;
 			}
 		});
-
-		////////////////////////////////////////////////////////////// Delete Class
-		////////////////////////////////////////////////////////////// Button
+		
+		////////////////////////////////////////////////////////////// Delete Class Button 
 		JButton btnDeleteClass = new JButton("Delete");
 		btnDeleteClass.setBounds(10, 165, 97, 21);
 		frmJambl.getContentPane().add(btnDeleteClass);
@@ -185,9 +183,9 @@ public class GUIView extends View {
 				saved = false;
 			}
 		});
-
-		////////////////////////////////////////////////////////////// Add Relationship
-		////////////////////////////////////////////////////////////// Button
+		
+		
+		////////////////////////////////////////////////////////////// Add Relationship Button 
 		JButton btnAddRelationship = new JButton("Add");
 		btnAddRelationship.setBounds(220, 105, 97, 21);
 		frmJambl.getContentPane().add(btnAddRelationship);
@@ -200,10 +198,8 @@ public class GUIView extends View {
 				saved = false;
 			}
 		});
-
-		////////////////////////////////////////////////////////////// Delete
-		////////////////////////////////////////////////////////////// Relationship
-		////////////////////////////////////////////////////////////// Button
+		
+		//////////////////////////////////////////////////////////////	Delete Relationship Button 
 		JButton btnDeleteRel = new JButton("Delete");
 		btnDeleteRel.setBounds(220, 136, 97, 21);
 		frmJambl.getContentPane().add(btnDeleteRel);
@@ -216,8 +212,8 @@ public class GUIView extends View {
 				saved = false;
 			}
 		});
-
-		////////////////////////////////////////////////////////////// Save Button
+		
+		//////////////////////////////////////////////////////////////	Save Button 
 		JButton btnSave = new JButton("Save");
 		btnSave.addActionListener(new ActionListener() {
 			@SuppressWarnings("static-access")
@@ -230,7 +226,7 @@ public class GUIView extends View {
 		});
 		btnSave.setBounds(10, 412, 85, 21);
 		frmJambl.getContentPane().add(btnSave);
-
+		
 		////////////////////////////////////////////////////////////// Load Button
 		JButton btnLoad = new JButton("Load");
 		btnLoad.setBounds(10, 445, 85, 21);
@@ -243,10 +239,9 @@ public class GUIView extends View {
 				loadW.initialize();
 			}
 		});
-
-		////////////////////////////////////////////////////////////// List All Button -
-		////////////////////////////////////////////////////////////// displays entire
-		////////////////////////////////////////////////////////////// UML diagram
+		
+		
+		////////////////////////////////////////////////////////////// List All Button - displays entire UML diagram
 		JButton btnListAll = new JButton("List All");
 		btnListAll.setBounds(10, 649, 137, 21);
 		frmJambl.getContentPane().add(btnListAll);
@@ -254,15 +249,11 @@ public class GUIView extends View {
 			@SuppressWarnings("static-access")
 			@Override
 			public void actionPerformed(ActionEvent e) {
-
+				
 			}
 		});
-
-		////////////////////////////////////////////////////////////// List
-		////////////////////////////////////////////////////////////// Relationships
-		////////////////////////////////////////////////////////////// Button - displays
-		////////////////////////////////////////////////////////////// list of
-		////////////////////////////////////////////////////////////// relationships
+		
+		////////////////////////////////////////////////////////////// List Relationships Button - displays list of relationships
 		JButton btnListRelationships = new JButton("List Relationships");
 		btnListRelationships.setBounds(10, 682, 137, 21);
 		frmJambl.getContentPane().add(btnListRelationships);
@@ -273,12 +264,8 @@ public class GUIView extends View {
 				textAreaMain.insert("Test", 0);
 			}
 		});
-
-		////////////////////////////////////////////////////////////// List a class
-		////////////////////////////////////////////////////////////// button - displays
-		////////////////////////////////////////////////////////////// a singular class
-		////////////////////////////////////////////////////////////// of a specified
-		////////////////////////////////////////////////////////////// name
+		
+		////////////////////////////////////////////////////////////// List a class button - displays a singular class of a specified name
 		JButton btnListClass = new JButton("List a Class");
 		btnListClass.setBounds(10, 616, 137, 21);
 		frmJambl.getContentPane().add(btnListClass);
@@ -290,117 +277,85 @@ public class GUIView extends View {
 				listCl.initialize();
 			}
 		});
-
-		////////////////////////////////////////////////////////////// Add Parameter
-		////////////////////////////////////////////////////////////// Button
+		
+		
+		////////////////////////////////////////////////////////////// Add Parameter Button
 		JButton btnAddParameter = new JButton("Add");
 		btnAddParameter.setBounds(10, 259, 97, 21);
 		frmJambl.getContentPane().add(btnAddParameter);
-
-		////////////////////////////////////////////////////////////// Remove Parameter
-		////////////////////////////////////////////////////////////// Button
+		
+		//////////////////////////////////////////////////////////////	Remove Parameter Button
 		JButton btnDeleteParameter = new JButton("Remove");
 		btnDeleteParameter.setBounds(10, 292, 97, 21);
 		frmJambl.getContentPane().add(btnDeleteParameter);
-
-		////////////////////////////////////////////////////////////// Change Parameter
-		////////////////////////////////////////////////////////////// Button
+		
+		////////////////////////////////////////////////////////////// Change Parameter Button
 		JButton btnChangeParameter = new JButton("Change");
 		btnChangeParameter.setBounds(10, 325, 97, 21);
 		frmJambl.getContentPane().add(btnChangeParameter);
-
+		
 		////////////////////////////////////////////////////////////// Add Method Button
 		JButton btnAddMethod = new JButton("Add");
 		btnAddMethod.setBounds(220, 256, 97, 21);
 		frmJambl.getContentPane().add(btnAddMethod);
-		btnAddMethod.addActionListener(new ActionListener() {
-			@SuppressWarnings("static-access")
-			public void actionPerformed(ActionEvent e) {
-				AddMethodW addMethodWindow = new AddMethodW();
-				addMethodWindow.initialize();
-			}
-		});
-
-		////////////////////////////////////////////////////////////// Change Method
-		////////////////////////////////////////////////////////////// Button
+		
+		////////////////////////////////////////////////////////////// Change Method Button
 		JButton btnRefactor = new JButton("Change");
 		btnRefactor.setBounds(220, 289, 97, 21);
 		frmJambl.getContentPane().add(btnRefactor);
-
-		////////////////////////////////////////////////////////////// Delete Method
-		////////////////////////////////////////////////////////////// Button
+		
+		////////////////////////////////////////////////////////////// Delete Method Button
 		JButton btnDeleteMethod = new JButton("Delete");
-		btnDeleteMethod.setBounds(220, 355, 97, 21);
-		frmJambl.getContentPane().add(btnDeleteMethod);
 		btnDeleteMethod.addActionListener(new ActionListener() {
-			@SuppressWarnings("static-access")
 			public void actionPerformed(ActionEvent e) {
-				DeleteMethodW deleteMethodWindow = new DeleteMethodW();
-				deleteMethodWindow.initialize();
 			}
 		});
-
-		////////////////////////////////////////////////////////////// Rename Method
-		////////////////////////////////////////////////////////////// Button
+		btnDeleteMethod.setBounds(220, 355, 97, 21);
+		frmJambl.getContentPane().add(btnDeleteMethod);
+		
+		////////////////////////////////////////////////////////////// Rename Method Button
 		JButton btnRenameMethod = new JButton("Rename");
 		btnRenameMethod.setBounds(220, 322, 97, 21);
 		frmJambl.getContentPane().add(btnRenameMethod);
-		btnRenameMethod.addActionListener(new ActionListener() {
-			@SuppressWarnings("static-access")
-			public void actionPerformed(ActionEvent e) {
-				RenameMethodW rnmMthd = new RenameMethodW();
-				rnmMthd.initialize();
-			}
-		});
-
+		
 		////////////////////////////////////////////////////////////// Add Field Button
 		JButton btnAddField = new JButton("Add");
 		btnAddField.setBounds(220, 437, 97, 21);
 		frmJambl.getContentPane().add(btnAddField);
-		btnAddField.addActionListener(new ActionListener() {
-			@SuppressWarnings("static-access")
-			public void actionPerformed(ActionEvent e) {
-				AddFieldW addFld = new AddFieldW();
-				addFld.initialize();
-			}
-		});
-
-		////////////////////////////////////////////////////////////// Change Field Type
-		////////////////////////////////////////////////////////////// Button
+		
+		////////////////////////////////////////////////////////////// Change Field Type Button
 		JButton btnChangeFieldType = new JButton("Edit Type");
 		btnChangeFieldType.setBounds(220, 470, 97, 21);
 		frmJambl.getContentPane().add(btnChangeFieldType);
-
-		////////////////////////////////////////////////////////////// Delete Field
-		////////////////////////////////////////////////////////////// Button
+		
+		////////////////////////////////////////////////////////////// Delete Field Button
 		JButton btnDeleteField = new JButton("Delete");
 		btnDeleteField.setBounds(220, 536, 97, 21);
 		frmJambl.getContentPane().add(btnDeleteField);
-
-		////////////////////////////////////////////////////////////// Rename Field
-		////////////////////////////////////////////////////////////// Button
+		
+		////////////////////////////////////////////////////////////// Rename Field Button
 		JButton btnRenameField = new JButton("Rename");
 		btnRenameField.setBounds(220, 503, 97, 21);
 		frmJambl.getContentPane().add(btnRenameField);
-		btnRenameField.addActionListener(new ActionListener() {
+		btnListClass.addActionListener(new ActionListener() {
 			@SuppressWarnings("static-access")
 			@Override
 			public void actionPerformed(ActionEvent e) {
-
+				
 			}
 		});
-
+		
 		frmJambl.addWindowListener(new WindowAdapter() {
 			@SuppressWarnings("static-access")
 			@Override
-			public void windowClosing(WindowEvent windowEvent) {
-				SaveAssuranceW savedCheck = new SaveAssuranceW();
-				savedCheck.initialize();
-			}
+			public void windowClosing(WindowEvent windowEvent){
+	            SaveAssuranceW savedCheck = new SaveAssuranceW();
+	            savedCheck.initialize();
+	         } 
 		});
 	}
-
+	
 	public void updateTextAreaMain() {
-
+		
 	}
 }
