+++ Application/src/PopUps/ListAClassWindow.java
@@ -16,7 +16,7 @@ import javax.swing.JButton;
 import javax.swing.JComboBox;
 import javax.swing.JFrame;
 import javax.swing.JLabel;
-import source.*;
+
 public class ListAClassWindow {
 
 	/**
