+++ Application/src/PopUps/LoadWindow.java
@@ -9,7 +9,7 @@ package PopUps;
 import java.awt.EventQueue;
 
 import javax.swing.JFrame;
-import source.*;
+
 public class LoadWindow {
 
 	/**
