--- Application/src/PopUps/RenameMethodW.java
@@ -1,23 +1,0 @@
-package PopUps;
-
-import javax.swing.JFrame;
-import source.*;
-public class RenameMethodW {
-	/**
-	 * Launch the application.
-	 */
-	public static void main(String[] args) {
-		initialize();
-	}
-
-	/**
-	 * Initialize the contents of the frame.
-	 */
-	public static void initialize() {
-		JFrame frame = new JFrame("JAMBL - Rename Method");
-		frame.setBounds(100, 100, 450, 300);
-		frame.setDefaultCloseOperation(JFrame.HIDE_ON_CLOSE);
-		frame.getContentPane().setLayout(null);
-		frame.setVisible(true);
-	}
-}
