+++ Application/src/PopUps/SaveAssuranceW.java
@@ -8,7 +8,7 @@
 package PopUps;
 import java.awt.EventQueue;
 import javax.swing.JFrame;
-import source.*;
+
 public class SaveAssuranceW {
 
 	/**
